fix rating counts on statistics, movie module, etc (no longer is rating divided by 2), cleanup and simplify more stuff now that there is no tech to consider

+++ TASVideos/Pages/Publications/Models/PublicationDisplayModel.cs
@@ -33,7 +33,7 @@ public class PublicationDisplayModel
 	public IEnumerable<FileModel> MovieFileLinks => Files
 		.Where(f => f.Type == FileType.MovieFile);
 
-	public double RatingCount { get; set; }
+	public int RatingCount { get; set; }
 	public double? OverallRating { get; set; }
 
 	public class TagModel
